timeline storage error
Match-id-25143f1553afc0cee6c8ace751f2f14f2bad9c1c

+++ analysis/ms_interface/msprof_data_storage.py
@@ -187,7 +187,7 @@ class MsprofDataStorage:
         :return: None
         """
         self.data_list = data_list
-        self.data_list.sort(key=lambda x: int(x.get('ts', 0)))
+        self.data_list.sort(key=lambda x: float(x.get('ts', 0)))
         self.set_tid()
         self._update_timeline_head()
